+++ src/components/button-group/index.js
@@ -1,0 +1,6 @@
+import ButtonGroup from '../button/src/button-group';
+ButtonGroup.install = function (Vue) {
+  Vue.component(ButtonGroup.name,ButtonGroup);
+};
+
+export default ButtonGroup;