+++ src/components/button/src/button-group.vue
@@ -1,0 +1,17 @@
+<template>
+    <div class="w-button-group">
+        <slot></slot>
+    </div>
+</template>
+
+<script>
+  export default {
+    name: "WButton-group"
+  }
+</script>
+
+<style lang="scss" scoped>
+    .w-button-group{
+        display: inline-block;
+    }
+</style>